Make warp bin part of the engine
alongside a few small bugfixes, the warp bin is now a part of the engine and can (hopefully) be used anywhere!
I'm not super confident that this won't eventually break, so just in case it does, the old warp bin implementation still exists in the main hub mod, albeit with the files renamed so they don't get used.