Music Logos
Like the one you see in the field of hopes and dreams and then never again. It's a modified version of Bor's library, made to use a font instead of premade sprites. An example of it in use is found in the Susie introduction cutscene.
Additionally, the music used in said cutscene ("Get on the Bus" from the Earthbound OST) has been updated to loop properly.